requirements.txt
Django
psycopg2-binary
djangorestframework
PyYAML
uritemplate